databases and spec files for all models up and running

+++ db/migrate/20150208015513_create_ratings.rb
@@ -2,7 +2,7 @@ class CreateRatings < ActiveRecord::Migration
   def change
     create_table :ratings do |t|
       t.float :total
-      t.float :out_of
+      t.integer :out_of
       t.string :unit
 
       t.timestamps null: false
